// src/core/interfaces/base/emailTemplate.interface.ts
export interface EmailTemplateInterface {
    user: any;
    mailTitle: string;
    mailContent: string;
    urlAction: string;
    buttonActionTitle: string;
    buttonBackgroundColor: any;
    buttonColor: any;
    greatingWord: string;
    footerAllRightReservedText: string;
    allRightReservedYears: number;
    appName: string;
    conditionUsingText: string;
    politicsText: string;
    protocol: string;
}